# packages/contracts/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/site-builder/sites/{id}/assets/presign": {
      "post": {
        "operationId": "AssetsController_presign_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PresignAssetResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "UUID 或请求体格式错误",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "NOT_FOUND",
                            "ASSET_VALIDATION_FAILED",
                            "ASSET_UPLOAD_INCOMPLETE",
                            "ASSET_DUPLICATE",
                            "ASSET_STATE_CONFLICT",
                            "ASSET_BUSY",
                            "ASSET_STORAGE_UNAVAILABLE",
                            "ASSET_COMMIT_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "当前 workspace 不可见该 Site",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "NOT_FOUND",
                            "ASSET_VALIDATION_FAILED",
                            "ASSET_UPLOAD_INCOMPLETE",
                            "ASSET_DUPLICATE",
                            "ASSET_STATE_CONFLICT",
                            "ASSET_BUSY",
                            "ASSET_STORAGE_UNAVAILABLE",
                            "ASSET_COMMIT_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "素材类型、MIME 或大小不合格",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "NOT_FOUND",
                            "ASSET_VALIDATION_FAILED",
                            "ASSET_UPLOAD_INCOMPLETE",
                            "ASSET_DUPLICATE",
                            "ASSET_STATE_CONFLICT",
                            "ASSET_BUSY",
                            "ASSET_STORAGE_UNAVAILABLE",
                            "ASSET_COMMIT_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "502": {
            "description": "对象存储暂不可用",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "NOT_FOUND",
                            "ASSET_VALIDATION_FAILED",
                            "ASSET_UPLOAD_INCOMPLETE",
                            "ASSET_DUPLICATE",
                            "ASSET_STATE_CONFLICT",
                            "ASSET_BUSY",
                            "ASSET_STORAGE_UNAVAILABLE",
                            "ASSET_COMMIT_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "素材上传第 1 步：校验并签发直传 URL（15 分钟有效）",
        "tags": [
          "SiteBuilder"
        ],
        "x-required-scopes": [
          "acquisition:write"
        ]
      }
    },
    "/api/v1/site-builder/assets/{id}/commit": {
      "post": {
        "operationId": "AssetsController_commit_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CommitAssetResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Asset UUID 格式错误",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "NOT_FOUND",
                            "ASSET_VALIDATION_FAILED",
                            "ASSET_UPLOAD_INCOMPLETE",
                            "ASSET_DUPLICATE",
                            "ASSET_STATE_CONFLICT",
                            "ASSET_BUSY",
                            "ASSET_STORAGE_UNAVAILABLE",
                            "ASSET_COMMIT_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "当前 workspace 不可见该 Asset",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "NOT_FOUND",
                            "ASSET_VALIDATION_FAILED",
                            "ASSET_UPLOAD_INCOMPLETE",
                            "ASSET_DUPLICATE",
                            "ASSET_STATE_CONFLICT",
                            "ASSET_BUSY",
                            "ASSET_STORAGE_UNAVAILABLE",
                            "ASSET_COMMIT_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "未上传、重复、忙或状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "NOT_FOUND",
                            "ASSET_VALIDATION_FAILED",
                            "ASSET_UPLOAD_INCOMPLETE",
                            "ASSET_DUPLICATE",
                            "ASSET_STATE_CONFLICT",
                            "ASSET_BUSY",
                            "ASSET_STORAGE_UNAVAILABLE",
                            "ASSET_COMMIT_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "素材内容校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "NOT_FOUND",
                            "ASSET_VALIDATION_FAILED",
                            "ASSET_UPLOAD_INCOMPLETE",
                            "ASSET_DUPLICATE",
                            "ASSET_STATE_CONFLICT",
                            "ASSET_BUSY",
                            "ASSET_STORAGE_UNAVAILABLE",
                            "ASSET_COMMIT_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "502": {
            "description": "对象存储暂不可用",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "NOT_FOUND",
                            "ASSET_VALIDATION_FAILED",
                            "ASSET_UPLOAD_INCOMPLETE",
                            "ASSET_DUPLICATE",
                            "ASSET_STATE_CONFLICT",
                            "ASSET_BUSY",
                            "ASSET_STORAGE_UNAVAILABLE",
                            "ASSET_COMMIT_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "提交持久化暂不可用",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "NOT_FOUND",
                            "ASSET_VALIDATION_FAILED",
                            "ASSET_UPLOAD_INCOMPLETE",
                            "ASSET_DUPLICATE",
                            "ASSET_STATE_CONFLICT",
                            "ASSET_BUSY",
                            "ASSET_STORAGE_UNAVAILABLE",
                            "ASSET_COMMIT_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "素材上传第 3 步：魔数/大小/去重校验 → 归位；doc 类进 KB 队列",
        "tags": [
          "SiteBuilder"
        ],
        "x-required-scopes": [
          "acquisition:write"
        ]
      }
    },
...
  },
  "info": {
    "title": "Global API",
    "description": "出海企业 AI 全球客户开发与增长平台 · 后端 API（前端接入见 packages/contracts/INTEGRATION.md）",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "同源部署（相对路径；具体 host 由部署环境决定）"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PresignAssetResponseDto": {
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string",
            "format": "uuid"
          },
          "uploadUrl": {
            "type": "string",
            "format": "uri"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "assetId",
          "uploadUrl",
          "expiresAt"
        ]
      },
      "PresignDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "logo",
              "product_image",
              "factory_image",
              "cert",
              "doc",
              "video"
            ]
          },
          "filename": {
            "type": "string"
          },
          "size": {
            "type": "number",
            "description": "字节数"
          },
          "mime": {
            "type": "string",
            "example": "image/jpeg"
          }
        },
        "required": [
          "kind",
          "filename",
          "size",
          "mime"
        ]
      },
      "CommitAssetResponseDto": {
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string",
            "format": "uuid"
          },
          "processingStatus": {
            "type": "string",
            "enum": [
              "queued",
              "ready"
            ]
          }
        },
        "required": [
          "assetId",
          "processingStatus"
        ]
      },
...
    }
  }
}